Add support for authenticating using jwt token
This commit adds the apps which has the support for authentication using
the jwt token. The users app generates a jwt token and stores it in the
session cookies whenever a user logins and the cart fetches this token
from the cookies and use it for auth. It queries the cart items using the
user_id and returns the cart items of the user.